@@ -1,7 +1,7 @@
 package com.cleveroad.adaptivetablelayout;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import android.view.ViewGroup;
 
 /**
