@@ -1,8 +1,8 @@
 package com.cleveroad.adaptivetablelayout;
 
 import android.content.Context;
-import android.support.annotation.Nullable;
-import android.support.v4.view.GestureDetectorCompat;
+import androidx.annotation.Nullable;
+import androidx.core.view.GestureDetectorCompat;
 import android.view.GestureDetector;
 import android.view.MotionEvent;
 
